+++ app/views/Profile/ProfilePost.tsx
@@ -1,34 +1,50 @@
-import { View, Text, Pressable, Touchable } from 'react-native'
-import React from 'react'
-import { useNavigation } from 'expo-router'
-import { SafeAreaView } from 'react-native-safe-area-context'
-import VideoCard from '~/components/VideoCard/VideoCard'
+import { View, Text, Pressable } from 'react-native';
+import { useNavigation } from 'expo-router';
+import { SafeAreaView } from 'react-native-safe-area-context';
+import VideoCard from '~/components/VideoCard/VideoCard';
 
-const ProfilePost = ({route}) => {
+const ProfilePost = ({ route }) => {
   // const navigation = useNavigation();
-  const data = route.params?.data
+  const data = route.params?.data;
   const navigation = useNavigation();
   return (
-    <View style={{flexDirection: "column", height: "100%"}}>
-      <SafeAreaView style={{backgroundColor: "white"}}>
-
-        <View style={{flexDirection: "row", alignItems: "center", padding: 10}}>
-          <View style={{position: "absolute", height: "100%", justifyContent: "center", zIndex: 99}}>
-            <Pressable style={{paddingHorizontal: 20}} onPress={() => {navigation.goBack()}}>
-              <Text style={{color: "blue"}}>
-                Back
-              </Text>
+    <View style={{ flexDirection: 'column', height: '100%' }}>
+      <SafeAreaView style={{ backgroundColor: 'white' }}>
+        <View
+          style={{ flexDirection: 'row', alignItems: 'center', padding: 10 }}
+        >
+          <View
+            style={{
+              position: 'absolute',
+              height: '100%',
+              justifyContent: 'center',
+              zIndex: 99,
+            }}
+          >
+            <Pressable
+              style={{ paddingHorizontal: 20 }}
+              onPress={() => {
+                navigation.goBack();
+              }}
+            >
+              <Text style={{ color: 'blue' }}>Back</Text>
             </Pressable>
           </View>
 
-          <View style={{width: "100%", justifyContent: "center", alignItems: "center"}}>
-            <Text style={{fontSize: 18}}>{data.name}</Text>
+          <View
+            style={{
+              width: '100%',
+              justifyContent: 'center',
+              alignItems: 'center',
+            }}
+          >
+            <Text style={{ fontSize: 18 }}>{data.name}</Text>
             <Text>{data.username}</Text>
           </View>
         </View>
       </SafeAreaView>
 
-      <View style={{width: "100%", flexGrow: 1}}>
+      <View style={{ width: '100%', flexGrow: 1 }}>
         <VideoCard
           videoIndex={0}
           currViewableIndex={0}
@@ -36,11 +52,11 @@ const ProfilePost = ({route}) => {
         />
       </View>
 
-      <View style={{height: 100}}>
+      <View style={{ height: 100 }}>
         <Text>{data.description}</Text>
       </View>
     </View>
-  )
-}
+  );
+};
 
-export default ProfilePost
+export default ProfilePost;
